Refactor image navigation e2e tests
